+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 0.2.0 (unreleased)
 
+- Added `Dockerfile`.
+
 
 ## 0.1.0 (2023-01-29)