+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 0.2.0 (unreleased)
 
-- Added `Dockerfile`.
+- Added `Dockerfile` and Compose file.
 
 
 ## 0.1.0 (2023-01-29)