+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## 0.2.0 (unreleased)
+## 0.2.0 (2023-02-03)
 
 - Added support for Discord and Mattermost as destinations.